+++ spec/interactor/context_spec.rb
@@ -144,9 +144,11 @@ module Interactor
       end
 
       it "makes the context available from the failure" do
-        context.fail!
-      rescue Failure => error
-        expect(error.context).to eq(context)
+        begin
+          context.fail!
+        rescue Failure => error
+          expect(error.context).to eq(context)
+        end
       end
     end
 
@@ -283,6 +285,42 @@ module Interactor
 
       it "is empty by default" do
         expect(context._called).to eq([])
+      end
+    end
+
+    describe "#deconstruct_keys" do
+      let(:context) { Context.build(foo: :bar) }
+
+      let(:deconstructed) { context.deconstruct_keys([:foo, :success, :failure, :halted]) }
+
+      it "deconstructs as hash pattern" do
+        expect(deconstructed[:foo]).to eq(:bar)
+      end
+
+      it "includes success and failure" do
+        expect(deconstructed[:success]).to eq(true)
+        expect(deconstructed[:failure]).to eq(false)
+      end
+
+      it "includes halted as false by default" do
+        expect(deconstructed[:halted]).to eq(false)
+      end
+
+      context "when halted" do
+        before do
+          context.halt!
+        rescue Halt
+          nil
+        end
+
+        it "includes halted as true" do
+          expect(context.deconstruct_keys(nil)[:halted]).to eq(true)
+        end
+
+        it "supports rightward assignment for halted:" do
+          context => { halted: }
+          expect(halted).to be(true)
+        end
       end
     end
   end